Struct diem_swarm::swarm::DiemSwarm

pub struct DiemSwarm {
    pub dir: DiemSwarmDir,
    pub nodes: HashMap<String, DiemNode>,
    pub config: SwarmConfig,
    pub role: RoleType,
    // some fields omitted
}

Struct holding instances and information of Diem Swarm

Fields
dir: DiemSwarmDirnodes: HashMap<String, DiemNode>config: SwarmConfigrole: RoleType
Implementations

impl DiemSwarm

pub fn configure_fn_swarm(
diem_node_bin_path: &Path,
config_dir: Option<String>,
template: Option<NodeConfig>,
upstream_config: &SwarmConfig,
fn_type: FullnodeType
) -> Result<DiemSwarm>

pub fn configure_validator_swarm(
diem_node_bin_path: &Path,
num_nodes: usize,
config_dir: Option<String>,
template: Option<NodeConfig>
) -> Result<DiemSwarm>

pub fn launch(&mut self)

pub fn launch_attempt(&mut self) -> Result<(), SwarmLaunchFailure>

pub fn wait_for_all_nodes_to_catchup(&mut self) -> bool

This function first checks the last committed round of all the nodes, picks the max value and then waits for all the nodes to catch up to that round. Once done, we can guarantee that all the txns committed before the invocation of this function are now available at all the nodes.

pub fn get_client_port(&self, index: usize) -> u16

A specific public JSON RPC port of a validator or a full node.

pub fn get_validators_ids(&self) -> Vec<String>

Vector with the peer ids of the validators in the swarm.

pub fn get_validators_debug_ports(&self) -> Vec<u16>

Vector with the debug ports of all the validators in the swarm.

pub fn get_validator(&self, idx: usize) -> Option<&DiemNode>

pub fn kill_node(&mut self, idx: usize)

pub fn add_node(&mut self, idx: usize) -> Result<(), SwarmLaunchFailure>